+++ chain/client/src/test_utils.rs
@@ -1033,6 +1033,8 @@ impl TestEnv {
     pub fn process_block(&mut self, id: usize, block: Block, provenance: Provenance) {
         let (mut accepted_blocks, result) = self.clients[id].process_block(block, provenance);
         assert!(result.is_ok(), format!("{:?}", result));
+        //  TODO
+        /*
         let more_accepted_blocks = self.clients[id].run_catchup(&vec![]).unwrap();
         accepted_blocks.extend(more_accepted_blocks);
         for accepted_block in accepted_blocks {
@@ -1042,6 +1044,7 @@ impl TestEnv {
                 accepted_block.provenance,
             );
         }
+         */
     }
 
     /// Produces block by given client, which kicks of creation of chunk.
